Renamed delta-consumer-worship-services to --->
dispatcher-worship-mandates
