+++ config/delta/rules.js
@@ -15,7 +15,7 @@ import resource from './resource';
 import syncWithKalliopeErrorNotification from './sync-with-kalliope-error-notification';
 import toezichtFlattenedFormDataGenerator from './toezicht-flattened-form-data-generator';
 import validateSubmission from './validate-submission';
-import deltaConsumerWorshipMandates from './delta-consumer-worship-services';
+import dispatcherWorshipMandates from './dispatcher-worship-mandates';
 import vendorDataDistribution from './vendor-data-distribution';
 import virusScanner from './virus-scanner';
 
@@ -37,7 +37,7 @@ export default [
   ...syncWithKalliopeErrorNotification,
   ...toezichtFlattenedFormDataGenerator,
   ...validateSubmission,
-  ...deltaConsumerWorshipMandates,
+  ...dispatcherWorshipMandates,
   ...vendorDataDistribution,
   ...virusScanner
 ];
